♻️ refactor: 로그인/로그아웃 요청 함수 폴더 service -> api 변경
