+++ src/components/layout/UserProfile.tsx
@@ -1,4 +1,4 @@
-import { useLogoutMutation } from '@/pages/login/service/postLogout';
+import { useLogoutMutation } from '@/pages/login/api/postLogout';
 import Popover from '../common/Popover';
 import router from 'next/router';
 import { ProfileImage } from '@/pages/mypage/components/profile/ProfileImage';
@@ -7,7 +7,7 @@ import useMemberStore from '@/stores/useMemberStore';
 export default function UserProfile({ nickname }: { nickname: string }) {
   const { mutate: logoutMutation } = useLogoutMutation();
   const { user } = useMemberStore();
-  
+
   const popoverItems = [
     {
       id: 'mypage',
@@ -23,13 +23,14 @@ export default function UserProfile({ nickname }: { nickname: string }) {
     },
   ];
 
-
-
   return (
     <div className="flex items-center">
       <div className="flex items-center gap-3">
         <Popover items={popoverItems} type="user">
-          <button type="button" className="mt-2 w-[40px] h-[40px] rounded-full overflow-hidden focus:outline-none">
+          <button
+            type="button"
+            className="mt-2 w-[40px] h-[40px] rounded-full overflow-hidden focus:outline-none"
+          >
             <ProfileImage imageUrl={user.profileImageUrl} size={40} />
           </button>
         </Popover>
